Add `kubernetes` and `docker-compose` sub-commands
These sub-commands are pretty much the same functionality as the `containerimage yaml`
sub-command. The main idea is that they are easier to figure out and type. The kubernetes
sub-command takes the "image regex" argument, which allows one to use an alternative
key to fetch a container image from a manifest. The `docker-compose` sub-command doesn't;
It'll just look for the `image` key in the YAML and attempt to replace that.