Allow excluding just the action name
Frizbee allowed to exclude the whole action name including the tag e.g.
`actions/checkout@v4`. But this is not user-friendly for cases where the
exclude would be upgraded over time as the exclude would have to be kept
in sync.

Let's also support the case where the exclude is just in the form of
`owner/action_name`.