Resolve commit ref correctly.
The command `frizbee actions <action>@<tag>` incorrectly retrieves the
tag's ref rather than the commit's one. The tag's ref should instead
be used to retrieve the details of the tag, which contain the ref of
the commit.

This change changes `getCheckSumForTag` to lookup the right field
using two subsequent call to GitHub.

Fixes #206